requirements.txt
streamlit
python-dotenv
requests
faiss-cpu
sentence-transformers
numpy
PyPDF2
tqdm
scikit-learn
langchain
